Add warning about required resource providers
We recently went through the conversion process in this document, found that the resource providers were not registered, and ultimately had a failure.  We had to get Microsoft support involved and had to rebuild the CMG.  To save others from this failure, this should be referenced in the documentation and a link given to the new providers.